More compiler warnings addressed (#5414).  Mostly the same as was in the 1.11 branch with r27741 r27739 r27729 r27723
git-svn-id: https://svn.osgeo.org/gdal/trunk@27745 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/grib/degrib18/g2clib-1.0.4/g2_addfield.c
@@ -345,7 +345,7 @@ g2int g2_addfield(unsigned char *cgrib,g2int ipdsnum,g2int *ipdstmpl,
               width=ndpts;
               height=1;
            }
-           else if ( width==allones || height==allones ) {
+           else if ( (unsigned int)width==allones || (unsigned int)height==allones ) {
               width=ndpts;
               height=1;
            }
